requirements.txt
tinydb
python-dateutil
colorama
pyserial
